push code

+++ payme-sdk/src/main/java/vn/payme/sdk/model/InfoPayment.kt
@@ -4,11 +4,11 @@ class InfoPayment {
     var action : String? = null
     var amount : Int? = null
     var note : String? = null
-    var orderId : Int? = null
-    var storeId : Int? = null
+    var orderId : Long? = null
+    var storeId : Long? = null
     var type : String? = null
 
-    constructor(action : String?,amount : Int?,note : String?,orderId : Int?,storeId : Int?,type : String?){
+    constructor(action : String?,amount : Int?,note : String?,orderId : Long?,storeId : Long?,type : String?){
         this.action = action
         this.amount = amount
         this.note = note
